test-bridge/requirements.txt
kafka-python
requests
streamlit
matplotlib
pandas